Cleanup
Expose some types at the crate root level, remove unnecessary folders, …
