+++ src/builtin/filter/slugify.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use serde_json::Value;
 use slug;
 
-use crate::engine::context::Context;
+use crate::context::Context;
 use crate::error::Result;
 
 pub(crate) fn slugify(value: &Value, _args: &HashMap<String, Value>, _context: &Context) -> Result<Value> {
