+++ src/builtin/function/uuidv4.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use serde_json::Value;
 use uuid::Uuid;
 
-use crate::engine::context::Context;
+use crate::context::Context;
 use crate::error::Result;
 
 pub(crate) fn uuidv4(_args: &HashMap<String, Value>, _context: &Context) -> Result<Value> {
@@ -16,7 +16,7 @@ mod tests {
 
     use uuid::{Uuid, Version};
 
-    use crate::engine::context::Context;
+    use crate::context::Context;
 
     use super::uuidv4;
