+++ tests/engine/eval/math.rs
@@ -1,7 +1,6 @@
 use serde_json::json;
 
-use balena_temen::engine::context::Context;
-use balena_temen::engine::Engine;
+use balena_temen::{Context, Engine};
 
 macro_rules! test_eval_eq {
     ($e:expr, $r:expr) => {{
